UI: Rename Feeds Manager -> Job Distributor
Feeds Manager is an outdated term, now that we are rolling out (Job Distributor)[https://github.com/smartcontractkit/job-distributor], we want to start adopting the new name.

We are only performing the rename on a superficial level which is on the UI, the rest including the backend remains untouched to keep things simple.

JIRA: https://smartcontract-it.atlassian.net/browse/OPCORE-891