fix(chain): override apollo ID matching for chain
If the chains query returns different chain with same id eg
```
[
                {
                    "id": "1",
                    "enabled": true,
                    "network": "aptos",
                    "__typename": "Chain"
                },
                {
                    "id": "1",
                    "enabled": true,
                    "network": "evm",
                    "__typename": "Chain"
                }
]
```

The default apollo client behaviour is to assume ID is always unique, but it is not the case here so things start to become confused like in the screen shot where both “1” show up for APTOS even though the UI correctly filters it.

The fix is to tell apollo client how to handle the uniqueness of the chain object.

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1384